Handle null profile number in updateNode.
When updateNode is called, it first checks to see if there is a node
in the database by address. If it finds one, it attempts to update it.

However, it is possible to have the node in the database but have the
profileNum field be null.  This causes Polyglot to get into a state
where it thinks everything is up-to-date, but in reality, the node
doesn't really exist.  This is really bad if it's the controller node
for the node server.

The fix is fairly simple.  If the node exist in the database, check
to see if the profile number is null. If it is, treat it the same as
if the node doesn't exist.  Polyglot will then go through the steps
to create the node and put everyting into a "normal" state.